new posts appear on top of page, users can delete posts from their profile, deleting updates state, image_urls now have user who posted username attached to the end to make suer cloud does not delete another users posts of the same url

+++ client/src/Post.tsx
@@ -2,30 +2,41 @@ import React, { useContext } from "react"
 import { storage } from './firebase'
 import { ref, uploadBytes, listAll, getDownloadURL, getStorage, deleteObject } from 'firebase/storage'
 import { ImageListContext } from "./Context"
-import { LoggedInUserContext } from "./Context"
+import { LoggedInUserContext, LoggedInUserPostsContext, PostsContext } from "./Context"
 
 
 function Post({url}:{url:string}) {
 
     const {setImageList, imageList} = useContext(ImageListContext)
     const {loggedInUser} = useContext(LoggedInUserContext)
+    const {setLoggedInUserPosts} = useContext(LoggedInUserPostsContext)
+    const {posts, setPosts} = useContext(PostsContext)
     const imageListRef = ref(storage, 'images/')
     const imageRef = ref(storage, url)
+    
     console.log(url)
     function handleDeletePost() {
-        loggedInUser.posts.forEach((post:any) => {
-            if (post.image_url === url) {
-                console.log(imageRef)
-                console.log(imageList)
+        console.log('h')
+        console.log(url)
+        posts.forEach((post:any) => {
+            if (post.image_url === url && post.user_id === loggedInUser.id) {
                 listAll(imageListRef)
                 .then((resp) => resp.items.forEach((item) => {
                     console.log(item)
                     getDownloadURL(item)  
                     .then((url) => {
-                        console.log(url)
-                        setImageList((prev:string[]) => prev.filter((img) => {
+                        const updatedList = imageList.filter((img:any) => {
                             return img !== url
-                        }))
+                        })
+                        const updatedPosts = posts.filter((post:any) => {
+                            return post.image_url !== url
+                        })
+                        setImageList(updatedList)
+                        setPosts(updatedPosts)
+                        console.log(imageList)
+                        fetch(`/users/${loggedInUser.id}/posts/${post.id}`, {
+                            method: 'DELETE'
+                        })
                     })
                     .then(() => deleteObject(imageRef))
                 }
@@ -38,9 +49,9 @@ function Post({url}:{url:string}) {
     }
 
     return(
-        <div className="col">
+        <div >
             <img className="post-pic" src={url} alt='oops'/>
-            <button >Delete!</button>
+            <button onClick={handleDeletePost}>Delete!</button>
         </div>
     )
 }
